requirements.txt
Flask~=3.1.1
flask
flask-cors
Flask-SQLAlchemy
Flask-Migrate~=4.1.0
Flask-Login~=0.6.3
Flask-WTF
python-dotenv~=1.1.0
PyMySQL
Werkzeug~=3.1.3
Flask-JWT-Extended~=4.7.1
alembic~=1.16.1
SQLAlchemy~=2.0.41